Edit and delete icons changed Cards for activity view Code cleaned

+++ src/app/activities/myActivities.component.ts
@@ -4,10 +4,12 @@ import {Router} from '@angular/router';
 import {MdDialog} from '@angular/material';
 import {NewActivityComponent} from './newActivity.component';
 import {UserService} from '../services/user.service';
+import {DialogConfirmationComponent} from "./dialogConfirmation.component";
 
 @Component({
   selector: 'app-my-activities',
   templateUrl: './myActivities.component.html',
+  styleUrls: ['myActivities.component.scss']
 })
 
 export class MyActivitiesComponent {
@@ -21,19 +23,19 @@ export class MyActivitiesComponent {
 
   load_activity(activity_id) {
     this.activityService.load_activity(activity_id).then( res => {
-      this.router.navigate(['activity_apps']);
+      this.router.navigate(['activity_apps/' + activity_id]);
     });
   }
 
   show_activity(activity_id) {
     this.activityService.load_activity(activity_id).then(res => {
-      this.router.navigate(['activity_view']);
+      this.router.navigate(['activity_view/' + activity_id]);
     });
   }
 
   edit_activity(activity_id) {
     this.activityService.load_activity(activity_id).then( res => {
-      this.router.navigate(['activity_edit']);
+      this.router.navigate(['activity_edit/' + activity_id]);
     });
   }
 
@@ -43,7 +45,14 @@ export class MyActivitiesComponent {
   }
 
   delete_activity(activityId) {
-    this.activityService.delete_activity(activityId);
+    const dialogRef = this.dialog.open(DialogConfirmationComponent, {
+      data: { message: 'Voulez vous vraiment supprimer cette activité ?' },
+    });
+    dialogRef.componentInstance.dialogRef = dialogRef;
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {this.activityService.delete_activity(activityId); }
+    });
+    //this.activityService.delete_activity(activityId);
   }
 
   duplicate_activity(activityId) {
